+++ .gitignore
@@ -1,2 +1,4 @@
 # Vim swap files
 *.swp
+
+gitsign
