+++ internal/git/git.go
@@ -11,9 +11,9 @@ import (
 	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/pkg/errors"
 	"github.com/sigstore/cosign/cmd/cosign/cli/fulcio/fulcioroots"
-	"github.com/sigstore/gitsign/fulcio"
-	"github.com/sigstore/gitsign/rekor"
-	"github.com/sigstore/gitsign/signature"
+	"github.com/sigstore/gitsign/internal/fulcio"
+	"github.com/sigstore/gitsign/internal/rekor"
+	"github.com/sigstore/gitsign/internal/signature"
 	"github.com/sigstore/rekor/pkg/generated/models"
 )
